[INCR-6]: Moved to new package - basics
