+++ src/main/java/com/bootcamp/spring/basics/controller/HelloWorldController.java
@@ -1,4 +1,4 @@
-package com.bootcamp.spring.controller;
+package com.bootcamp.spring.basics.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
